adiciona variavel global email
agora tem confirmaçao de excluir conta para o idoso

+++ app/src/main/java/com/example/idosologin/menuIdoso.kt
@@ -6,9 +6,11 @@ import android.util.Log
 import android.widget.Button
 import android.widget.EditText
 import android.widget.Toast
+import androidx.appcompat.app.AlertDialog
 import androidx.appcompat.app.AppCompatActivity
 import com.google.firebase.auth.FirebaseAuth
 import com.google.firebase.auth.FirebaseUser
+import com.google.firebase.firestore.FirebaseFirestore
 
 class menuIdoso : AppCompatActivity() {
 
@@ -17,6 +19,10 @@ class menuIdoso : AppCompatActivity() {
     private lateinit var updateEmailButton: Button
     private lateinit var deleteAccountButton: Button
     private lateinit var logoutButton: Button
+
+    val db = FirebaseFirestore.getInstance()
+
+
 
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
@@ -30,6 +36,7 @@ class menuIdoso : AppCompatActivity() {
         logoutButton = findViewById(R.id.btnLogout2)
 
 
+
         loadUserInfo()
 
 
@@ -39,7 +46,27 @@ class menuIdoso : AppCompatActivity() {
 
 
         deleteAccountButton.setOnClickListener {
-            deleteAccount()
+            // Create an AlertDialog Builder
+            val builder = AlertDialog.Builder(this)
+            builder.setTitle("Confirmação")
+            builder.setMessage("Tem certeza que quer deletar a conta?")
+
+            // Set positive button (Yes)
+            builder.setPositiveButton("Sim") { dialog, _ ->
+                deleteAccount()
+                db.collection("idoso").document("${MainActivity.GlobalData.ultEmail}")
+                    .delete()
+                dialog.dismiss()
+            }
+
+            // Set negative button (No)
+            builder.setNegativeButton("Não") { dialog, _ ->
+                // Do nothing, just dismiss the dialog
+                dialog.dismiss()
+            }
+
+            // Show the dialog
+            builder.show()
         }
